+++ native/GhostRigger.Tools.Workflow.ModuleMeshes/Python/src/gui/panels/module_editor/__init__.py
@@ -5,6 +5,7 @@ from .module_editor_asset_browser import ModuleEditorAssetBrowser
 from .module_editor_properties import ModuleEditorPropertiesPanel
 from .module_editor_toolbar import ModuleEditorToolbar
 from .module_editor_viewport_panel import ModuleEditorViewportPanel
+from .readiness_panel import ModuleReadinessPanel
 from .validation_panel import ModuleValidationPanel
 
 __all__ = [
@@ -13,5 +14,6 @@ __all__ = [
     "ModuleEditorPropertiesPanel",
     "ModuleEditorToolbar",
     "ModuleEditorViewportPanel",
+    "ModuleReadinessPanel",
     "ModuleValidationPanel",
 ]